Remove token and id state and use new credentials prop
I broadened the prop being passed through components to be a credentials
prop that includes both token and id. I also realized that in order for
the sign out button to successfully transition back to the login page,
appContext always needs to be passed in as a prop.
QuestPage: just turning token and id into credentials
QuestionScreen: token and id into credentials and add appContext to navBar
component props
UploadScreen: same as above