Create UpdateQuestionScreen
This is modeled after the create question form. The differences should be
that the text fields come populated with the previous data, passed through
props to the state. Upon changing those fields, the states change. When
the form is submitted, the data passed to the back end is passed through
the updated states.