Add LocationComponent
This makes the processing of location cards so much easier. Instead of
making a click handler for each card, now there is a LocationComponent
which maps through the location titles and images to create a individual
card with click handler for each location. The LocationComponent runs an
index request to the back end and filters out only the questions with the
card's location.